Updates include: - (edit) add a suffix 'Section' to components in pages/home/* and pages/about/*(except Heros) and move to components/Home/* and components/About/* respectively - export Home/* and About/* components using its corresponding index.js
