+++ src/components/About/index.js
@@ -1,0 +1,3 @@
+export { AboutCCDLSection } from './AboutCCDLSection'
+export { AboutOverviewSection } from './AboutOverviewSection'
+export { AboutHero } from './AboutHero'
