+++ src/components/Home/index.js
@@ -1,0 +1,5 @@
+export { HomeDocumentsSection } from './HomeDocumentsSection'
+export { HomeExamplesSection } from './HomeExamplesSection'
+export { HomeFeaturesSection } from './HomeFeaturesSection'
+export { HomeSignUpSection } from './HomeSignUpSection'
+export { HomeHero } from './HomeHero'
